Garmin: use COORDINATE base field for Weather coordinate fields (untested)
Use the standard field encoding for coordinates. Untested as my watch does not show the contents of the coordinate fields, but verified that they used to be invalid in the test cases and are now properly populated.

Also update test cases to match.

+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/devices/garmin/GarminSupport.java
@@ -715,8 +715,8 @@ public class GarminSupport extends AbstractBTLESingleDeviceSupport implements IC
         today.setWindSpeed(weather.getWindSpeed());
         today.setTemperatureFeelsLike(weather.getFeelsLikeTemp());
         today.setRelativeHumidity(weather.getCurrentHumidity());
-        today.setObservedLocationLat((long) weather.getLatitude());
-        today.setObservedLocationLong((long) weather.getLongitude());
+        today.setObservedLocationLat((double) weather.getLatitude());
+        today.setObservedLocationLong((double) weather.getLongitude());
         today.setAirQuality(null); //ensure the definition is added
         if (null != weather.getAirQuality()) {
             today.setAirQuality(FieldDefinitionWeatherAqi.aqiAbsoluteValueToEnum(weather.getAirQuality().getAqi()));
